clean (#881)
