--- .circleci/config.yml
@@ -1,109 +1,0 @@
-version: 2
-
-jobs:
-  test-3.7: &test-template
-    docker:
-      - image: circleci/python:3.7
-
-    working_directory: ~/repo
-
-    steps:
-      - checkout
-
-      # Download and cache dependencies
-      - restore_cache:
-          keys:
-          - v1-dependencies-{{ checksum "requirements.txt" }}
-          # fallback to using the latest cache if no exact match is found
-          - v1-dependencies-
-
-      - run:
-          name: Install dependencies
-          command: |
-            python3 -m venv venv
-            venv/bin/pip install -U pip
-            venv/bin/pip install -r requirements.txt
-
-      - save_cache:
-          paths:
-            - ./venv
-          key: v1-dependencies-{{ checksum "requirements.txt" }}
-
-      - run:
-          name: Run tests
-          command: |
-            . venv/bin/activate
-            pip install codecov codacy-coverage
-            coverage run -m unittest discover .
-            codecov
-
-  test-3.6:
-    <<: *test-template
-    docker:
-      - image: circleci/python:3.6
-
-  test-3.5:
-    <<: *test-template
-    docker:
-      - image: circleci/python:3.5
-
-  test-2.7:
-    <<: *test-template
-    docker:
-      - image: circleci/python:2.7
-
-
-  deploy:
-    working_directory: ~/repo
-    docker:
-      - image: circleci/python:3.7
-    steps:
-      - checkout
-
-      - run:
-          name: Setup env
-          command: |
-            python3 -m venv venv
-            venv/bin/pip install -U pip twine wheel
-            venv/bin/pip install -r requirements.txt
-
-      - run:
-          name: Upload to pypi
-          command: |
-            . venv/bin/activate
-            python setup.py verify
-            python setup.py sdist bdist_wheel
-            twine upload dist/* --repository-url https://test.pypi.org/legacy/
-            twine upload dist/*
-
-workflows:
-  version: 2
-  test-deploy:
-    jobs:
-      - test-3.7:
-          filters:
-            tags:
-              only: /^v.*/
-      - test-3.6:
-          filters:
-            tags:
-              only: /^v.*/
-      - test-3.5:
-          filters:
-            tags:
-              only: /^v.*/
-      - test-2.7:
-          filters:
-            tags:
-              only: /^v.*/
-      - deploy:
-          requires:
-            - test-3.7
-            - test-3.6
-            - test-3.5
-            - test-2.7
-          filters:
-            tags:
-              only: /^v.*/
-            branches:
-              ignore: /.*/
