+++ NEMbox/__init__.py
@@ -26,3 +26,14 @@ ________     __________________________  _____ _     _
 + ------------------------------------------ +
 
 """
+from .utils import create_dir
+from .utils import create_file
+from .const import Constant
+
+
+def create_config():
+    create_dir(Constant.conf_dir)
+    create_dir(Constant.download_dir)
+    create_file(Constant.storage_path)
+    create_file(Constant.log_path, default="")
+    create_file(Constant.cookie_path, default="#LWP-Cookies-2.0\n")
