--- create_venv.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-python3 -m venv venv && venv/bin/pip install -U pip && venv/bin/pip install -r requirements-dev.txt
