--- requirements.txt
@@ -1,5 +1,0 @@
-requests
-requests-cache
-pycryptodomex
-future
-fuzzywuzzy
